fix(function): fix register order of unary minus.
